requirements.txt
psycopg2-binary==2.9.11
pandas==3.0.0
numpy==2.4.2
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
xgboost>=2.0.0
prophet>=1.1
pmdarima>=2.0.0
statsmodels>=0.14.0
tensorflow>=2.13.0
joblib>=1.3.0
tqdm>=4.65.0
